# payment-app/src/main/resources/api.yaml
openapi: 3.0.9
info:
  title: Payment API
  contact:
    email: thevivek2@gmail.com
  version: 1.0.0

paths:
  /api/v1/mobile/accounts:
    post:
      tags:
        - Account
      summary: Create Account
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResource'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoilationResource'

  /api/v1/accounts/{accountNumber}:
    get:
      tags:
        - Account
      summary: Get Account Detail
      operationId: getAccount
      parameters:
        - name: accountNumber
          in: path
          description: Account Number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountResource'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoilationResource'
    patch:
      tags:
        - Account
      summary: Partial Update of Account
      operationId: patchAccount
      parameters:
        - name: accountNumber
          in: path
          description: Account Number
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResource'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoilationResource'

  /api/v1/accounts/{accountNumber}/history:
    get:
      tags:
        - Account
      summary: Get Account History
      operationId: getAccountHistory
      parameters:
        - name: accountNumber
          in: path
          description: Account Number
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/AccountResource'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoilationResource'

  /api/v1/payments:
    post:
      tags:
        - Payment
      summary: Make Payment Transfer
      operationId: createPayment
      requestBody:
       content:
         application/json:
           schema:
             $ref: '#/components/schemas/PaymentRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoilationResource'


components:
  schemas:
    AccountRequest:
      properties:
        mobileNumber:
          type: string
          example: '2222222222'
          minLength: 10
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: 'NRS'
          default: 'NRS'
      required:
        - mobileNumber


    VoilationResource:
      properties:
        field:
          type: string
          example: 'email'
        message:
          type: string
          example: 'This is required'
      required:
        - message

    AccountResource:
      properties:
        accountNumber:
          type: string
          example: '22222'
        status:
          $ref: '#/components/schemas/AccountStatus'

    AccountStatus:
      type: string
      enum:
        - OPEN
        - CLOSED

    AccountUpdateRequest:
      properties:
        status:
          $ref: '#/components/schemas/AccountStatus'

    PaymentRequest:
      properties:
        debtorAccountNumber:
          type: string
          example: '22222222'
        creditorAccountNumber:
          type: string
          example: '3333333'
        amount:
          type: number
          format: bigdecimal
          example: '2000'

    PaymentResponse:
      properties:
        status:
          type: string
          example: 'SAVED'